Moxed tests & implemented tast_to_ast
