+++ lib/src/tests/mod.rs
@@ -1,3 +1,11 @@
+mod arithmetic_methods_class;
+mod assigned_fields_class;
+mod bool_alg_class;
+mod complex_if_class;
+mod empty_class;
+mod empty_method_class;
+mod fib_class;
+mod if_class;
 mod to_java;
 
 use self::to_java::class_to_java;
@@ -18,13 +26,81 @@ fn normalize_str(s: std::string::String) -> std::string::String {
         .fold("".to_string(), |acc, s| acc + s)
 }
 
-fn create_test_file(class: &Class, name: &str) {
+fn stmt_tast_to_ast(stmt: &Stmt) -> Stmt {
+    match stmt {
+        Stmt::TypedStmt(x, typ) => stmt_tast_to_ast(x),
+        Stmt::Block(stmts) => Block(stmts.iter().map(|x| stmt_tast_to_ast(x)).collect()),
+        Stmt::Return(expr) => Return(expr_tast_to_ast(expr)),
+        Stmt::While(cond, body) => While(expr_tast_to_ast(cond), Box::new(stmt_tast_to_ast(body))),
+        Stmt::If(cond, body, elze) => If(
+            expr_tast_to_ast(cond),
+            Box::new(stmt_tast_to_ast(body)),
+            match elze {
+                Some(x) => Some(Box::new(stmt_tast_to_ast(x))),
+                None => None,
+            },
+        ),
+        Stmt::StmtExprStmt(stmt_expr) => StmtExprStmt(stmt_expr_tast_to_ast(stmt_expr)),
+        default => stmt.clone(),
+    }
+}
+
+fn stmt_expr_tast_to_ast(stmt_expr: &StmtExpr) -> StmtExpr {
+    match stmt_expr {
+        StmtExpr::Assign(var, val) => Assign(var.clone(), expr_tast_to_ast(val)),
+        StmtExpr::New(typ, params) => New(
+            typ.clone(),
+            params.iter().map(|x| expr_tast_to_ast(x)).collect(),
+        ),
+        StmtExpr::MethodCall(obj, method, params) => MethodCall(
+            expr_tast_to_ast(obj),
+            method.clone(),
+            params.iter().map(|x| expr_tast_to_ast(x)).collect(),
+        ),
+        StmtExpr::TypedStmtExpr(x, typ) => stmt_expr_tast_to_ast(x),
+    }
+}
+
+fn expr_tast_to_ast(expr: &Expr) -> Expr {
+    match expr {
+        Expr::InstVar(x, s) => InstVar(Box::new(expr_tast_to_ast(x)), s.clone()),
+        Expr::Unary(s, x) => Unary(s.clone(), Box::new(expr_tast_to_ast(x))),
+        Expr::Binary(op, l, r) => Binary(
+            op.clone(),
+            Box::new(expr_tast_to_ast(l)),
+            Box::new(expr_tast_to_ast(r)),
+        ),
+        Expr::StmtExprExpr(x) => StmtExprExpr(Box::new(stmt_expr_tast_to_ast(x))),
+        Expr::TypedExpr(x, t) => expr_tast_to_ast(x),
+        default => expr.clone(),
+    }
+}
+
+fn tast_to_ast(class: &Class) -> Class {
+    Class {
+        name: class.name.clone(),
+        fields: class.fields.clone(),
+        methods: class
+            .methods
+            .clone()
+            .into_iter()
+            .map(|method| MethodDecl {
+                ret_type: method.ret_type.clone(),
+                name: method.name.clone(),
+                params: method.params.clone(),
+                body: stmt_tast_to_ast(&method.body),
+            })
+            .collect(),
+    }
+}
+
+fn create_test_file(ast: &Class, tast: Option<&Class>, name: &str) {
     let file_path = format!("tests/{name}.java");
     let gen_file_path = format!("tests/{name}-gen.java");
     let class_file_path = format!("tests/{name}.class");
 
     // Generate Java Code from AST and write to file
-    let class_code = class_to_java(class);
+    let class_code = class_to_java(ast);
     let mut file =
         File::create(gen_file_path.clone()).expect("File for generated code couldn't be created");
     file.write(class_code.as_bytes())
@@ -79,450 +155,71 @@ fn create_test_file(class: &Class, name: &str) {
     // Write AST & TAST to files
     let mut file =
         File::create(format!("tests/{name}-AST.json")).expect("File couldn't be created");
-    serde_json::to_writer_pretty(&mut file, &class).expect("Couldn't serialize class");
+    serde_json::to_writer_pretty(&mut file, &ast).expect("Couldn't serialize class");
 
-    // let mut file =
-    //     File::create(format!("tests/{name}-TAST.json")).expect("File couldn't be created");
-    // serde_json::to_writer_pretty(&mut file, &class).expect("Couldn't serialize class");
+    if let Some(tast) = tast {
+        let mut file =
+            File::create(format!("tests/{name}-TAST.json")).expect("File couldn't be created");
+        serde_json::to_writer_pretty(&mut file, tast).expect("Couldn't serialize class");
+    }
 }
 
-#[test]
-fn if_class() {
-    let class = Class {
-        name: "If".to_string(),
-        fields: vec![],
-        methods: vec![MethodDecl {
-            name: "f".to_string(),
-            params: vec![(Type::Char, "c".to_string())],
-            ret_type: Type::Bool,
-            body: Block(vec![
-                If(
-                    Binary(
-                        "==".to_string(),
-                        Box::new(LocalOrFieldVar("c".to_string())),
-                        Box::new(Char('a')),
-                    ),
-                    Box::new(Return(Bool(true))),
-                    None,
-                ),
-                Return(Bool(false)),
-            ]),
-        }],
-    };
-    create_test_file(&class, "If");
-}
-
-#[test]
-fn arithmetic_methods_class() {
-    let class = Class {
-        name: "ArithmeticMethods".to_string(),
-        fields: vec![
-            FieldDecl {
-                field_type: Type::Int,
-                name: "x".to_string(),
-                val: Some("69".to_string()),
-            },
-            FieldDecl {
-                field_type: Type::Int,
-                name: "y".to_string(),
-                val: Some("420".to_string()),
-            },
-        ],
-        methods: vec![
-            MethodDecl {
-                ret_type: Type::Int,
-                name: "addX".to_string(),
-                params: vec![(Type::Int, "a".to_string())],
-                body: Block(vec![Return(Binary(
-                    "+".to_string(),
-                    Box::new(LocalOrFieldVar("x".to_string())),
-                    Box::new(LocalOrFieldVar("a".to_string())),
-                ))]),
-            },
-            MethodDecl {
-                ret_type: Type::Int,
-                name: "addY".to_string(),
-                params: vec![(Type::Int, "a".to_string())],
-                body: Block(vec![Return(Binary(
-                    "+".to_string(),
-                    Box::new(LocalOrFieldVar("y".to_string())),
-                    Box::new(LocalOrFieldVar("a".to_string())),
-                ))]),
-            },
-            MethodDecl {
-                ret_type: Type::Int,
-                name: "complexMath".to_string(),
-                params: vec![(Type::Int, "a".to_string()), (Type::Int, "b".to_string())],
-                body: Block(vec![
-                    StmtExprStmt(Assign(
-                        "a".to_string(),
-                        Binary(
-                            "*".to_string(),
-                            Box::new(LocalOrFieldVar("y".to_string())),
-                            Box::new(Binary(
-                                "/".to_string(),
-                                Box::new(Binary(
-                                    "+".to_string(),
-                                    Box::new(LocalOrFieldVar("a".to_string())),
-                                    Box::new(LocalOrFieldVar("b".to_string())),
-                                )),
-                                Box::new(LocalOrFieldVar("x".to_string())),
-                            )),
-                        ),
-                    )),
-                    StmtExprStmt(Assign(
-                        "b".to_string(),
-                        Binary(
-                            "+".to_string(),
-                            Box::new(LocalOrFieldVar("a".to_string())),
-                            Box::new(Unary(
-                                "-".to_string(),
-                                Box::new(LocalOrFieldVar("b".to_string())),
-                            )),
-                        ),
-                    )),
-                    StmtExprStmt(Assign(
-                        "a".to_string(),
-                        Binary(
-                            "+".to_string(),
-                            Box::new(LocalOrFieldVar("x".to_string())),
-                            Box::new(Binary(
-                                "*".to_string(),
-                                Box::new(LocalOrFieldVar("b".to_string())),
-                                Box::new(LocalOrFieldVar("a".to_string())),
-                            )),
-                        ),
-                    )),
-                    Return(Binary(
-                        "*".to_string(),
-                        Box::new(LocalOrFieldVar("x".to_string())),
-                        Box::new(LocalOrFieldVar("a".to_string())),
-                    )),
-                ]),
-            },
-        ],
-    };
-    create_test_file(&class, "ArithmeticMethods");
-}
-
-#[test]
-fn assigned_fields_class() {
-    let class = Class {
-        name: "AssignedFields".to_string(),
-        fields: vec![
-            FieldDecl {
-                field_type: Type::Int,
-                name: "x".to_string(),
-                val: Some("69".to_string()),
-            },
-            FieldDecl {
-                field_type: Type::Char,
-                name: "c".to_string(),
-                val: Some("x".to_string()),
-            },
-            FieldDecl {
-                field_type: Type::String,
-                name: "s".to_string(),
-                val: Some("Hello World".to_string()),
-            },
-            FieldDecl {
-                field_type: Type::String,
-                name: "stringsCanBeNull".to_string(),
-                val: Some("null".to_string()),
-            },
-            FieldDecl {
-                field_type: Type::Bool,
-                name: "b".to_string(),
-                val: Some("true".to_string()),
-            },
-        ],
-        methods: vec![],
-    };
-    create_test_file(&class, "AssignedFields");
-}
-
-#[test]
-fn bool_alg_class() {
-    let class = Class {
-        name: "BoolAlg".to_string(),
-        fields: vec![],
-        methods: vec![MethodDecl {
-            ret_type: Type::Bool,
-            name: "f".to_string(),
-            params: vec![
-                (Type::Bool, "a".to_string()),
-                (Type::Bool, "b".to_string()),
-                (Type::Bool, "c".to_string()),
-            ],
-            body: Block(vec![Return(Binary(
-                "||".to_string(),
-                Box::new(Binary(
-                    "&&".to_string(),
-                    Box::new(Binary(
-                        "&&".to_string(),
-                        Box::new(LocalOrFieldVar("a".to_string())),
-                        Box::new(LocalOrFieldVar("b".to_string())),
-                    )),
-                    Box::new(LocalOrFieldVar("c".to_string())),
-                )),
-                Box::new(LocalOrFieldVar("c".to_string())),
-            ))]),
-        }],
-    };
-    create_test_file(&class, "BoolAlg");
-}
-
-#[test]
-fn complex_if_class() {
-    let class = Class {
-        name: "ComplexIf".to_string(),
-        fields: vec![],
-        methods: vec![MethodDecl {
-            ret_type: Type::Bool,
-            name: "f".to_string(),
-            params: vec![(Type::Char, "c".to_string())],
-            body: Block(vec![If(
-                Binary(
-                    "==".to_string(),
-                    Box::new(LocalOrFieldVar("c".to_string())),
-                    Box::new(Expr::Char('a')),
-                ),
-                Box::new(Block(vec![Return(Expr::Bool(true))])),
-                Some(Box::new(If(
-                    Binary(
-                        "==".to_string(),
-                        Box::new(LocalOrFieldVar("c".to_string())),
-                        Box::new(Expr::Char('b')),
-                    ),
-                    Box::new(Block(vec![Return(Expr::Bool(false))])),
-                    Some(Box::new(If(
-                        Binary(
-                            "==".to_string(),
-                            Box::new(LocalOrFieldVar("c".to_string())),
-                            Box::new(Expr::Char('c')),
-                        ),
-                        Box::new(Return(Expr::Bool(true))),
-                        Some(Box::new(Block(vec![If(
-                            Binary(
-                                "||".to_string(),
-                                Box::new(Binary(
-                                    "==".to_string(),
-                                    Box::new(LocalOrFieldVar("c".to_string())),
-                                    Box::new(Expr::Char('d')),
-                                )),
-                                Box::new(Binary(
-                                    "==".to_string(),
-                                    Box::new(LocalOrFieldVar("c".to_string())),
-                                    Box::new(Expr::Char('e')),
-                                )),
-                            ),
-                            Box::new(Return(Expr::Bool(false))),
-                            Some(Box::new(If(
-                                Binary(
-                                    "||".to_string(),
-                                    Box::new(Binary(
-                                        "==".to_string(),
-                                        Box::new(LocalOrFieldVar("c".to_string())),
-                                        Box::new(Expr::Char('d')),
-                                    )),
-                                    Box::new(Binary(
-                                        "==".to_string(),
-                                        Box::new(LocalOrFieldVar("c".to_string())),
-                                        Box::new(Expr::Char('e')),
-                                    )),
-                                ),
-                                Box::new(Return(Expr::Bool(true))),
-                                Some(Box::new(Block(vec![Return(Expr::Bool(false))]))),
-                            ))),
-                        )]))),
-                    ))),
-                ))),
-            )]),
-        }],
-    };
-    create_test_file(&class, "ComplexIf");
-}
-
-#[test]
-fn empty_class() {
-    let class = Class {
-        name: "Empty".to_string(),
-        fields: vec![],
-        methods: vec![],
-    };
-    create_test_file(&class, "Empty");
-}
-
-#[test]
-fn empty_method_class() {
-    let class = Class {
-        name: "emptyMethod".to_string(),
-        fields: vec![],
-        methods: vec![MethodDecl {
-            ret_type: Type::Void,
-            name: "f".to_string(),
-            params: vec![],
-            body: Block(vec![]),
-        }],
-    };
-    create_test_file(&class, "emptyMethod");
-}
-
-#[test]
-fn fib_class() {
-    let class = Class {
-        name: "Fib".to_string(),
-        fields: vec![],
-        methods: vec![
-            MethodDecl {
-                ret_type: Type::Int,
-                name: "rec".to_string(),
-                params: vec![(Type::Int, "n".to_string())],
-                body: Block(vec![If(
-                    Binary(
-                        "<".to_string(),
-                        Box::new(LocalOrFieldVar("n".to_string())),
-                        Box::new(Expr::Integer(2)),
-                    ),
-                    Box::new(Return(LocalOrFieldVar("n".to_string()))),
-                    Some(Box::new(Block(vec![Return(Binary(
-                        "+".to_string(),
-                        Box::new(StmtExprExpr(Box::new(MethodCall(
-                            Expr::This,
-                            "rec".to_string(),
-                            vec![Binary(
-                                "-".to_string(),
-                                Box::new(LocalOrFieldVar("n".to_string())),
-                                Box::new(Expr::Integer(1)),
-                            )],
-                        )))),
-                        Box::new(StmtExprExpr(Box::new(MethodCall(
-                            Expr::This,
-                            "rec".to_string(),
-                            vec![Binary(
-                                "-".to_string(),
-                                Box::new(LocalOrFieldVar("n".to_string())),
-                                Box::new(Expr::Integer(2)),
-                            )],
-                        )))),
-                    ))]))),
-                )]),
-            },
-            MethodDecl {
-                ret_type: Type::Int,
-                name: "iter".to_string(),
-                params: vec![(Type::Int, "n".to_string())],
-                body: Block(vec![
-                    If(
-                        Binary(
-                            "<".to_string(),
-                            Box::new(LocalOrFieldVar("n".to_string())),
-                            Box::new(Expr::Integer(2)),
-                        ),
-                        Box::new(Return(LocalOrFieldVar("n".to_string()))),
-                        None,
-                    ),
-                    LocalVarDecl(Type::Int, "x".to_string()),
-                    StmtExprStmt(Assign("x".to_string(), Expr::Integer(0))),
-                    LocalVarDecl(Type::Int, "y".to_string()),
-                    StmtExprStmt(Assign("y".to_string(), Expr::Integer(1))),
-                    LocalVarDecl(Type::Int, "i".to_string()),
-                    StmtExprStmt(Assign("i".to_string(), Expr::Integer(1))),
-                    While(
-                        Binary(
-                            "<".to_string(),
-                            Box::new(LocalOrFieldVar("i".to_string())),
-                            Box::new(LocalOrFieldVar("n".to_string())),
-                        ),
-                        Box::new(Block(vec![
-                            LocalVarDecl(Type::Int, "next".to_string()),
-                            StmtExprStmt(Assign(
-                                "next".to_string(),
-                                Binary(
-                                    "+".to_string(),
-                                    Box::new(LocalOrFieldVar("y".to_string())),
-                                    Box::new(LocalOrFieldVar("x".to_string())),
-                                ),
-                            )),
-                            StmtExprStmt(Assign("x".to_string(), LocalOrFieldVar("y".to_string()))),
-                            StmtExprStmt(Assign(
-                                "y".to_string(),
-                                LocalOrFieldVar("next".to_string()),
-                            )),
-                            StmtExprStmt(Assign(
-                                "i".to_string(),
-                                Binary(
-                                    "+".to_string(),
-                                    Box::new(LocalOrFieldVar("i".to_string())),
-                                    Box::new(Expr::Integer(1)),
-                                ),
-                            )),
-                        ])),
-                    ),
-                    Return(LocalOrFieldVar("y".to_string())),
-                ]),
-            },
-        ],
-    };
-    create_test_file(&class, "Fib");
-}
-
-// #[test]
+// use super::*\n#[test]
 // fn Fields_class() {
 //     let class = Class {};
-//     create_test_file(&class, "Fields");
+//     create_test_file(&tast_to_ast(&class), Some(&class), "Fields");
 // }
 
-// #[test]
+// use super::*\n#[test]
 // fn IntFields_class() {
 //     let class = Class {};
-//     create_test_file(&class, "IntFields");
+//     create_test_file(&tast_to_ast(&class), Some(&class), "IntFields");
 // }
 
-// #[test]
+// use super::*\n#[test]
 // fn LocalVarDecl_class() {
 //     let class = Class {};
-//     create_test_file(&class, "LocalVarDecl");
+//     create_test_file(&tast_to_ast(&class), Some(&class), "LocalVarDecl");
 // }
 
-// #[test]
+// use super::*\n#[test]
 // fn MethodCall_class() {
 //     let class = Class {};
-//     create_test_file(&class, "MethodCall");
+//     create_test_file(&tast_to_ast(&class), Some(&class), "MethodCall");
 // }
 
-// #[test]
+// use super::*\n#[test]
 // fn NamingConflict_class() {
 //     let class = Class {};
-//     create_test_file(&class, "NamingConflict");
+//     create_test_file(&tast_to_ast(&class), Some(&class), "NamingConflict");
 // }
 
-// #[test]
+// use super::*\n#[test]
 // fn Negator_class() {
 //     let class = Class {};
-//     create_test_file(&class, "Negator");
+//     create_test_file(&tast_to_ast(&class), Some(&class), "Negator");
 // }
 
-// #[test]
+// use super::*\n#[test]
 // fn Return_class() {
 //     let class = Class {};
-//     create_test_file(&class, "Return");
+//     create_test_file(&tast_to_ast(&class), Some(&class), "Return");
 // }
 
-// #[test]
+// use super::*\n#[test]
 // fn SetterGetter_class() {
 //     let class = Class {};
-//     create_test_file(&class, "SetterGetter");
+//     create_test_file(&tast_to_ast(&class), Some(&class), "SetterGetter");
 // }
 
-// #[test]
+// use super::*\n#[test]
 // fn StrAdd_class() {
 //     let class = Class {};
-//     create_test_file(&class, "StrAdd");
+//     create_test_file(&tast_to_ast(&class), Some(&class), "StrAdd");
 // }
 
-// #[test]
+// use super::*\n#[test]
 // fn While_class() {
 //     let class = Class {};
-//     create_test_file(&class, "While");
+//     create_test_file(&tast_to_ast(&class), Some(&class), "While");
 // }
